+++ main.go
@@ -37,7 +37,7 @@ func main() {
 		Title:            "Lucid Player",
 		Width:            1024,
 		Height:           768,
-		WindowStartState: options.Maximised,
+		WindowStartState: options.Normal,
 		AssetServer: &assetserver.Options{
 			Assets: assets,
 		},